run_pipeline.py
import argparse
import yaml
import os
import sys

def run_view_synthesis(config, input_image_path, output_dir):
    print(f"Simulating View Synthesis for {input_image_path} using config: {str(config)}")
    # Placeholder for actual view synthesis logic
    # This would generate view_01.png, view_02.png, ... and camera_poses.json
    os.makedirs(output_dir, exist_ok=True)
    with open(os.path.join(output_dir, "camera_poses.json"), "w") as f:
        f.write("{\"poses\": []}")
    print(f"View Synthesis output to {output_dir}")
    return output_dir

def run_3d_reconstruction(config, input_dir, output_dir):
    print(f"Simulating 3D Reconstruction from {input_dir} using config: {str(config)}")
    # Placeholder for actual 3D reconstruction logic (e.g., 3D Gaussian Splatting)
    os.makedirs(output_dir, exist_ok=True)
    with open(os.path.join(output_dir, "model.glb"), "w") as f:
        f.write("GLB_MODEL_DATA")
    print(f"3D Reconstruction output to {output_dir}")
    return output_dir

def run_rendering(config, input_model_path, output_dir):
    print(f"Simulating 2D Rendering of {input_model_path} using config: {str(config)}")
    # Placeholder for actual rendering logic (e.g., Blender script)
    os.makedirs(output_dir, exist_ok=True)
    with open(os.path.join(output_dir, "rendered.png"), "w") as f:
        f.write("RENDERED_IMAGE_DATA")
    print(f"2D Rendering output to {output_dir}")
    return output_dir

def package_results(config, input_dirs, output_deployment_dir):
    print(f"Simulating Packaging Results to {output_deployment_dir} from {input_dirs} using config: {str(config)}")
    os.makedirs(output_deployment_dir, exist_ok=True)
    
    # Create a dummy index.html
    with open(os.path.join(output_deployment_dir, "index.html"), "w") as f:
        f.write("<html><body><h1>PoC Results</h1></body></html>")
    
    # Simulate copying results from track_a and track_b
    for track_name, track_dir in input_dirs.items():
        track_output_path = os.path.join(output_deployment_dir, track_name)
        os.makedirs(track_output_path, exist_ok=True)
        # Simulate copying model.glb and rendered.png
        with open(os.path.join(track_output_path, "model.glb"), "w") as f:
            f.write(f"{track_name.upper()}_GLB_MODEL_DATA")
        with open(os.path.join(track_output_path, "rendered.png"), "w") as f:
            f.write(f"{track_name.upper()}_RENDERED_IMAGE_DATA")
        # Simulate copying view_*.png
        for i in range(1, 3): # Just 2 dummy views for now
            with open(os.path.join(track_output_path, f"view_0{i}.png"), "w") as f:
                f.write(f"{track_name.upper()}_VIEW_0{i}_IMAGE_DATA")

    print(f"Results packaged to {output_deployment_dir}")
    return output_deployment_dir

def main():
    parser = argparse.ArgumentParser(description="Run the Manga to 3D PoC Pipeline.")
    parser.add_argument("--config", type=str, default="config.yml", help="Path to the configuration file.")
    parser.add_argument("--input_image", type=str, default="input/original.png", help="Path to the input original image.")
    parser.add_argument("--output_dir", type=str, default=None, help="Optional: Path to the output deployment directory. Overrides config.")
    
    args = parser.parse_args()

    if not os.path.exists(args.config):
        print(f"Error: Configuration file not found at {args.config}", file=sys.stderr)
        sys.exit(1)

    with open(args.config, 'r') as f:
        config = yaml.safe_load(f)

    print(f"Starting PoC Pipeline with config: {args.config}")

    # Define output directories
    temp_dir = config.get("pipeline_temp_dir", "temp")
    output_deployment_dir = args.output_dir if args.output_dir else config.get("output_deployment_dir", "output_for_deployment")
    
    # Ensure input directory exists for dummy image
    os.makedirs(os.path.dirname(args.input_image), exist_ok=True)
    if not os.path.exists(args.input_image):
        with open(args.input_image, "w") as f:
            f.write("DUMMY_ORIGINAL_IMAGE_DATA")
        print(f"Created dummy input image: {args.input_image}")

    # --- Step 1: View Synthesis ---
    print("\n--- Running View Synthesis ---")
    track_a_views_dir = os.path.join(temp_dir, "track_a_views")
    track_b_views_dir = os.path.join(temp_dir, "track_b_views")

    track_a_cfg = config.get("track_a_config", {})
    print(f"DEBUG: Type of track_a_cfg: {type(track_a_cfg)}")
    run_view_synthesis(track_a_cfg.copy(), args.input_image, track_a_views_dir)

    track_b_cfg = config.get("track_b_config", {})
    print(f"DEBUG: Type of track_b_cfg: {type(track_b_cfg)}")
    run_view_synthesis(track_b_cfg.copy(), args.input_image, track_b_views_dir)

    # --- Step 2: 3D Reconstruction ---
    print("\n--- Running 3D Reconstruction ---")
    track_a_3d_dir = os.path.join(temp_dir, "track_a_3d")
    track_b_3d_dir = os.path.join(temp_dir, "track_b_3d")

    reconstruction_config_a = config.get("reconstruction_config", {})
    run_3d_reconstruction(reconstruction_config_a.copy(), track_a_views_dir, track_a_3d_dir)

    reconstruction_config_b = config.get("reconstruction_config", {})
    run_3d_reconstruction(reconstruction_config_b.copy(), track_b_views_dir, track_b_3d_dir)

    # --- Step 3: 2D Rendering ---
    print("\n--- Running 2D Rendering ---")
    track_a_render_dir = os.path.join(temp_dir, "track_a_render")
    track_b_render_dir = os.path.join(temp_dir, "track_b_render")

    rendering_config_a = config.get("rendering_config", {})
    run_rendering(rendering_config_a.copy(), os.path.join(track_a_3d_dir, "model.glb"), track_a_render_dir)

    rendering_config_b = config.get("rendering_config", {})
    run_rendering(rendering_config_b.copy(), os.path.join(track_b_3d_dir, "model.glb"), track_b_render_dir)

    # --- Step 4: Packaging Results ---
    print("\n--- Packaging Results ---")
    packaging_config = config.get("packaging_config", {})
    package_results(
        packaging_config.copy(),
        {
            "track_a": track_a_render_dir, # In a real scenario, this would be a more complex aggregation
            "track_b": track_b_render_dir
        },
        output_deployment_dir
    )

    print("\nPoC Pipeline Finished Successfully!")

if __name__ == "__main__":
    main()
